ai provider embeds
